requirements.txt
fastapi[standard]>=0.110.0
sqlmodel>=0.0.16
chromadb>=0.4.24
openai>=1.14.0
python-dotenv>=1.0.1
pydantic>=2.6.4
tenacity>=8.2.3
twilio>=9.0.0
discord.py>=2.3.2
